initial commit for repository

+++ myorderapp/src/main/java/za/ac/cput/myorderapp/conf/Factory/CustomerFactory.java
@@ -8,7 +8,7 @@ import java.util.Map;
  * Created by student on 2015/04/25.
  */
 public class CustomerFactory {
-    public static Customer createCustomer(int id, Map<String, String> values){
+    public static Customer createCustomer(Long id, Map<String, String> values){
         Customer customer = new Customer.Builder(values.get("name"))
                 .id(id)
                 .phone(values.get("phone"))
